% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarySearch.R
\name{BinarySearch}
\alias{BinarySearch}
\title{A function that calculates the maximum shrinkage for a given sparse penalty
This function examines te maximum elements of the eigenvector
and determines the level of shrinkage possible to meet the sparse penalty}
\usage{
BinarySearch(argu, sumabs)
}
\arguments{
\item{argu}{This is the eigenvector to undergo l1 penalization}

\item{sumabs}{This is the sparse penalty}
}
\description{
A function that calculates the maximum shrinkage for a given sparse penalty
This function examines te maximum elements of the eigenvector
and determines the level of shrinkage possible to meet the sparse penalty
}
\examples{
BinarySearch(argu,sumabs)
}
\keyword{shrinkage}
