% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_metrics.R
\name{categorical_metrics}
\alias{categorical_metrics}
\title{This function calculates the performance metrics for each categorical response.}
\usage{
categorical_metrics(ytrain, ytest, y_hat, nresp, resp.names)
}
\arguments{
\item{ytrain}{training response dataset}

\item{ytest}{testing response dataset}

\item{y_hat}{fitted/predicted values}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function calculates the performance metrics for each categorical response.
}
\examples{
categorical_metrics(ytrain, ytest, y_hat, nresp, resp.names)
}
\keyword{categorical}
\keyword{metrics}
