% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CifManPlot.R
\name{CifManPlot}
\alias{CifManPlot}
\title{This function provides a manhattan plot unique to CIFTI data}
\usage{
CifManPlot(
  Zlist,
  sepAnalysis,
  nresp,
  man.thresh,
  resp.names,
  specific.groups = NULL,
  exclude.groups = NULL,
  special.colors = NULL
)
}
\arguments{
\item{Zlist}{contains the eigenevectors corresponding to the loadings}

\item{sepAnalysis}{tells whether the eigenvectors should be displayed separately}

\item{nresp}{number of response vectors present in datasets}

\item{man.thresh}{flag for whether this is a manhattan plot}

\item{resp.names}{gives list of names for each response}

\item{specific.groups}{allows certain groups to be selected}

\item{exclude.groups}{allows certain groups to be excluded}

\item{color}{scheme for plot}
}
\description{
This function provides a manhattan plot unique to CIFTI data
}
\examples{
CifManPlot(Zlist, sepAnalysis, nresp, man.thresh, resp.names, 
  specific.groups=NULL, exclude.groups=NULL, special.colors=NULL)
}
\keyword{CIFTI}
\keyword{manhattan}
\keyword{plot}
