% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_metrics.R
\name{continuous_metrics}
\alias{continuous_metrics}
\title{This function calculates the performance metrics for each continuous response.}
\usage{
continuous_metrics(ytrain, ytest, y_hat, nresp, resp.names)
}
\arguments{
\item{ytrain}{training response dataset}

\item{ytest}{testing response dataset}

\item{y_hat}{fitted/predicted values}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function calculates the performance metrics for each continuous response.
}
\examples{
continuous_metrics(ytrain, ytest, y_hat, nresp, resp.names)
}
\keyword{continuous}
\keyword{metrics}
