% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densplotSq.R
\name{densplotSq}
\alias{densplotSq}
\title{This function provides visualization for the responses}
\usage{
densplotSq(TrTs, nresp, resp.names)
}
\arguments{
\item{TrTs}{this contains the fitted/predicted responses}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}

\item{sepAnalysis}{tells whether the eigenvectors should be displayed separately}
}
\description{
This function provides visualization for the responses
}
\examples{
densplotSq(TrTs, nresp, resp.names)
}
\keyword{density}
\keyword{distance}
\keyword{fitted}
\keyword{plot}
\keyword{square}
