% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densplotyhat.R
\name{densplotyhat}
\alias{densplotyhat}
\title{This function provides visualization for the responses}
\usage{
densplotyhat(y_hat, nresp, resp.names)
}
\arguments{
\item{y_hat}{this contains the fitted/predicted responses}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function provides visualization for the responses
}
\examples{
densplotyhat(y_hat, nresp, resp.names)
}
\keyword{density}
\keyword{fitted}
\keyword{plot}
