% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DualSPC.R
\name{DualSPC}
\alias{DualSPC}
\title{This function computes supervised principal component analysis, but uses the
dual supervised PCA form.  This is useful when you have far more predictors
than observations.  This allows the matrix to be calculated based on
the lower dimensionality of the number of observations rather than
the high dimensionality of the predictors.  It obtains
the eigenevectors from the HSIC criterion and encodes the training and
testing data}
\usage{
DualSPC(
  ytrain,
  xtrain,
  ytest,
  xtest,
  ycont,
  ybinary,
  nresp,
  strictEV,
  nTopEvecs
)
}
\arguments{
\item{ytrain}{training response dataset}

\item{xtrain}{training predictor dataset}

\item{ytest}{testing response dataset}

\item{xtest}{testing predictor dataset}

\item{ycont}{flag indicates response data is continuous}

\item{ybinary}{flag indicates response data is binary}

\item{nresp}{number of response vectors present in datasets}

\item{strictEV}{flag whether to impose strict requiresments to accept number
of eigenevectors or to select a more appropriate number}

\item{nTopEvecs}{number of desired eigenevectors}
}
\description{
This function computes supervised principal component analysis, but uses the
dual supervised PCA form.  This is useful when you have far more predictors
than observations.  This allows the matrix to be calculated based on
the lower dimensionality of the number of observations rather than
the high dimensionality of the predictors.  It obtains
the eigenevectors from the HSIC criterion and encodes the training and
testing data
}
\examples{
DualSPC(ytrain, xtrain, ytest, xtest, ycont, ybinary, nresp, strictEV, nTopEvecs)
}
\keyword{analysis}
\keyword{component}
\keyword{dual}
\keyword{principal}
\keyword{supervised}
