% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvecPlot.R
\name{EvecPlot}
\alias{EvecPlot}
\title{This function provides visualization for the loadings of each eigenvector
provided to the function.}
\usage{
EvecPlot(Zlist, sepAnalysis, nresp, resp.names)
}
\arguments{
\item{Zlist}{this contains the eigenvectors}

\item{sepAnalysis}{tells whether the eigenvectors should be displayed separately}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function provides visualization for the loadings of each eigenvector
provided to the function.
}
\examples{
EvecPlot(Zlist, sepAnalysis, nresp, resp.names)
}
\keyword{eigenvector}
\keyword{plot}
