% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMbin.R
\name{GLMbin}
\alias{GLMbin}
\title{This function calculates predictions for each binary response using all
eigenvectors together.}
\usage{
GLMbin(ytrain, ztrain, ytest, ztest, nresp)
}
\arguments{
\item{ytrain}{training response dataset}

\item{ztrain}{encoded training data}

\item{ytest}{testing response dataset}

\item{ztest}{encoded testing data}

\item{nresp}{number of response vectors present in datasets}
}
\description{
This function calculates predictions for each binary response using all
eigenvectors together.
}
\examples{
GLMbin(ytrain, ztrain, ytest, ztest, nresp)
}
\keyword{GLM}
\keyword{binary}
