% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMbinSep.R
\name{GLMbinSep}
\alias{GLMbinSep}
\title{This function calculates binary predictions for binary response data.}
\usage{
GLMbinSep(ytrain, ytest, Z, nresp, resp.names)
}
\arguments{
\item{ytrain}{training response dataset}

\item{ytest}{testing response dataset}

\item{Z}{encoded pca responses}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function calculates binary predictions for binary response data.
}
\examples{
GLMbinSep(ytrain, ytest, Z, nresp, resp.names)
}
\keyword{GLM}
\keyword{binary}
\keyword{separate}
