% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMmulti.R
\name{GLMmulti}
\alias{GLMmulti}
\title{This function calculates predictions for each multinomial response using all
eigenvectors together.}
\usage{
GLMmulti(ytrain, ztrain, ytest, ztest, nresp, resp.names)
}
\arguments{
\item{ytrain}{training response dataset}

\item{ztrain}{encoded training data}

\item{ytest}{testing response dataset}

\item{ztest}{encoded testing data}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function calculates predictions for each multinomial response using all
eigenvectors together.
}
\examples{
GLMmulti(ytrain, ztrain, ytest, ztest, nresp, resp.names)
}
\keyword{GLM}
\keyword{multinomial}
