% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMmultiSep.R
\name{GLMmultiSep}
\alias{GLMmultiSep}
\title{This function calculates predictions for multinomial response data.}
\usage{
GLMmultiSep(ytrain, ytest, Z, nresp, resp.names)
}
\arguments{
\item{ytrain}{training response dataset}

\item{ytest}{testing response dataset}

\item{Z}{encoded pca responses}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function calculates predictions for multinomial response data.
}
\examples{
GLMmultiSep(ytrain, ytest, Z, nresp, resp.names)
}
\keyword{GLM}
\keyword{multinomial}
\keyword{separate}
