% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2n.R
\name{l2n}
\alias{l2n}
\title{A function calculates the l2 norm
This function calculates the l2n norm of a given vector "vec"
If the l2 norm is zero, this function then returns a value of 0.05
so as not to result in a division by zero}
\usage{
l2n(vec)
}
\arguments{
\item{vec}{Vector from which to calculate l2 norm}
}
\description{
A function calculates the l2 norm
This function calculates the l2n norm of a given vector "vec"
If the l2 norm is zero, this function then returns a value of 0.05
so as not to result in a division by zero
}
\examples{
l2n(vec)
}
\keyword{l2}
\keyword{norm}
