% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMcont.R
\name{LMcont}
\alias{LMcont}
\title{This function calculates yhat for each continuous using all
eigenvectors together.}
\usage{
LMcont(ytrain, ztrain, ytest, ztest, nresp, resp.names)
}
\arguments{
\item{ytrain}{training response dataset}

\item{ztrain}{encoded training data}

\item{ytest}{testing response dataset}

\item{ztest}{encoded testing data}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function calculates yhat for each continuous using all
eigenvectors together.
}
\examples{
LMcont(ytrain, ztrain, ytest, ztest, nresp, resp.names)
}
\keyword{continuous}
\keyword{linear}
\keyword{model}
