% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMcontSep.R
\name{LMcontSep}
\alias{LMcontSep}
\title{This function calculates yhat for each continuous response separately.}
\usage{
LMcontSep(ytrain, ytest, Z, nresp, resp.names)
}
\arguments{
\item{ytrain}{training response dataset}

\item{ytest}{testing response dataset}

\item{Z}{encoded pca responses}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function calculates yhat for each continuous response separately.
}
\examples{
LMcontSep(ytrain, ytest, Z, nresp, resp.names)
}
\keyword{continuous}
\keyword{linear}
\keyword{model}
\keyword{separate}
