% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoopSPC.R
\name{LoopSPC}
\alias{LoopSPC}
\title{This function cycles through each response and computes supervised principal
component analysis eigenvectors.  This is caused by doing separate analysis.}
\usage{
LoopSPC(ytrain, xtrain, ytest, xtest, ycont, nresp, strictEV, nTopEvecs)
}
\arguments{
\item{ytrain}{training response dataset}

\item{xtrain}{training predictor dataset}

\item{ytest}{testing response dataset}

\item{xtest}{testing predictor dataset}

\item{ycont}{flag indicates response data is continuous}

\item{nresp}{number of response vectors present in datasets}

\item{strictEV}{flag whether to impose strict requiresments to accept number
of eigenevectors or to select a more appropriate number}

\item{nTopEvecs}{number of desired eigenevectors}
}
\description{
This function cycles through each response and computes supervised principal
component analysis eigenvectors.  This is caused by doing separate analysis.
}
\examples{
LoopSPC(ytrain, xtrain, ytest, xtest, ycont, nresp, strictEV, nTopEvecs)
}
\keyword{analysis}
\keyword{component}
\keyword{individual}
\keyword{principal}
\keyword{response}
\keyword{supervised}
