% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoopSSPC.R
\name{LoopSSPC}
\alias{LoopSSPC}
\title{This function cycles through each response and computes supervised principal
component analysis eigenvectors.  This is caused by doing separate analysis.
This invokes the SSPCA function to compute sparse eigenevectors
for each response.}
\usage{
LoopSSPC(
  xtrain,
  ytrain,
  xtest,
  ycont,
  nresp,
  sumabsv = 4,
  niter = 20,
  K = 1,
  orth = TRUE,
  trace = TRUE,
  v = NULL,
  center = FALSE,
  cnames = NULL,
  vpos = FALSE,
  vneg = FALSE,
  compute.pve = TRUE,
  strictEV = TRUE
)
}
\arguments{
\item{xtrain}{training predictor dataset}

\item{ytrain}{training response dataset}

\item{xtest}{testing predictor dataset}

\item{ycont}{flag indicates response data is continuous}

\item{nresp}{number of response vectors present in datasets}

\item{sumabsv}{causes shrinkage to occur to the loadings}

\item{niter}{the number of iterations allowed to achieve convergence}

\item{K}{number of eigenevectors allowed to be calculated}

\item{orth}{determines whether orthogonal vectors are required}

\item{trace}{this allows debugging}

\item{v}{focuses on the loadings vector}

\item{center}{this centers the predictor columns}

\item{cnames}{this gives column names}

\item{vpos}{this determines whether the matrix is to be positive only}

\item{vneg}{this determines whether the matrix is to be negative only}

\item{compute.pve}{computes cv using missing values}

\item{strictEV}{flag whether to impose strict requiresments to accept number
of eigenevectors or to select a more appropriate number}
}
\description{
This function cycles through each response and computes supervised principal
component analysis eigenvectors.  This is caused by doing separate analysis.
This invokes the SSPCA function to compute sparse eigenevectors
for each response.
}
\examples{
LoopSSPC(xtrain, ytrain, xtest, ycont, nresp, sumabsv=4, niter=20, 
  K=1, orth=TRUE, trace=TRUE, v=NULL, center=FALSE, 
  cnames=NULL, vpos=FALSE, vneg=FALSE, compute.pve=TRUE, 
  strictEV=TRUE)
}
\keyword{analysis}
\keyword{component}
\keyword{indiviual}
\keyword{principal}
\keyword{sparse}
\keyword{supervised}
