% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManPlot.R
\name{ManPlot}
\alias{ManPlot}
\title{This function provides basic manhattan plots for the loadings of each eigenvector
provided to the function.}
\usage{
ManPlot(Zlist, sepAnalysis, nresp, man.thresh, resp.names)
}
\arguments{
\item{Zlist}{this contains the eigenvectors}

\item{sepAnalysis}{tells whether the eigenvectors should be displayed separately}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function provides basic manhattan plots for the loadings of each eigenvector
provided to the function.
}
\examples{
ManPlot(Zlist, sepAnalysis, nresp, man.thresh, resp.names)
}
\keyword{manhattan}
\keyword{plot}
