% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix2Vector.R
\name{Matrix2Vector}
\alias{Matrix2Vector}
\title{This function converts pconn data matrix into a vector}
\usage{
Matrix2Vector(pconn_data = NULL, pconn_vector = NULL, direction = "to_vector")
}
\arguments{
\item{pconn_data}{accepts the pconn data matrix}

\item{pconn_vector}{accepts the pconn data in vector form}

\item{direction}{coded to convert matrix into vector}
}
\description{
This function converts pconn data matrix into a vector
}
\examples{
Matrix2Vector(pconn_data=NULL,pconn_vector=NULL,direction="to_vector")
}
\keyword{matrix}
\keyword{pconn}
\keyword{vector}
