% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.na.R
\name{mean.na}
\alias{mean.na}
\title{A function that calculates the mean value of all non-missing values
of a vector}
\usage{
\method{mean}{na}(vec)
}
\arguments{
\item{vec}{vector in which to calculate the mean of non-missing values}
}
\description{
A function that calculates the mean value of all non-missing values
of a vector
}
\examples{
mean.na(vec)
}
\keyword{mean}
\keyword{missing}
