% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSMDOrth.R
\name{MultiSMDOrth}
\alias{MultiSMDOrth}
\title{This function decomposes a matrix using singular value decomposition ad uses
the SMD function.  This results in vectors corresponding to the loadings
which have undergone shrinkage}
\usage{
MultiSMDOrth(x, sumabsu, sumabsv, niter, K, trace, v, vpos, vneg)
}
\arguments{
\item{x}{this is the matrix which is decomposed and the decomposition vectors
corresponding to the loadings are shrunk}

\item{sumabsu}{this parameter determines by how much the decomposition vectors
corresponding to the observations}

\item{sumabsv}{this parameter determines by how much the decomposition vectors
corresponding to the loadings}

\item{niter}{number of iterations to allow for convergence}

\item{K}{determines how many vectors should be calculated}

\item{trace}{option to deisplay potential warnings}

\item{v}{allows a particular v matrix to be passed to the SMD function}

\item{vpos}{allows the option to force the vector corresponding to
loadings to correspond to the parallel maxima using pmax}

\item{vneg}{allows the option to force the vector corresponding to
loadings to correspond to the parallel minima using pmin}

\item{upos}{allows the option to force the vector corresponding to
observations to correspond to the parallel maxima using pmax}

\item{uneg}{allows the option to force the vector corresponding to
observations to correspond to the parallel minima using pmin}
}
\description{
This function decomposes a matrix using singular value decomposition ad uses
the SMD function.  This results in vectors corresponding to the loadings
which have undergone shrinkage
}
\examples{
MultiSMDOrth(x,sumabsu,sumabsv,niter,K, trace, v, vpos, vneg)
}
\keyword{matrix}
\keyword{multiple}
\keyword{orthogonal}
\keyword{singular}
