% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmniSPCA.R
\name{OmniSPCA}
\alias{OmniSPCA}
\title{This function provides all options for analysis using SSPCA}
\usage{
OmniSPCA(
  xtrain,
  ytrain,
  xtest,
  ytest,
  resp.names = NULL,
  ycont = T,
  ybinary = F,
  sepAnalysis = F,
  K = 1,
  sumabsv = 0,
  strictEV = T,
  Sens.analysis = F,
  sepclass = T,
  binClassCombine = F,
  NoExtraEV = F,
  ydensity = F,
  dpyhat = F,
  dsq = F,
  qqplot = F,
  opplot = F,
  Evec.plot = F,
  manhattan.plot = F,
  cifti.man.plot = F,
  man.thresh = 0,
  niter = 20,
  orth = TRUE,
  trace = TRUE,
  v = NULL,
  center = TRUE,
  test_metrics = T,
  cnames = NULL,
  vpos = FALSE,
  vneg = FALSE,
  compute.pve = TRUE
)
}
\arguments{
\item{xtrain}{training predictor dataset}

\item{ytrain}{training response dataset}

\item{xtest}{testing predictor dataset}

\item{ytest}{testing response dataset}

\item{resp.names}{parameter containing response names}

\item{ycont}{flag indicates response data is continuous}

\item{ybinary}{flags for whether response is binary}

\item{sepAnalysis}{flag for whether analysis should be performed separately for
each response}

\item{K}{number of eigenevectors allowed to be calculated}

\item{sumabsv}{causes shrinkage to occur to the loadings}

\item{strictEV}{flag whether to impose strict requiresments to accept number
of eigenevectors or to select a more appropriate number}

\item{Sens.analysis}{not currently supported}

\item{binClassCombine}{not currently support}

\item{NoExtraEV}{unknown function}

\item{ydensity}{calculates density plot}

\item{dpyhat}{calculates y hat plot}

\item{dsq}{calculates squared error plot}

\item{qqplot}{generates qqplot}

\item{opplot}{calculates observed vs predicted plot}

\item{Evec.plot}{generates eigenevector plot}

\item{manhattan.plot}{generates basic manhattan plot}

\item{cifti.man.plot}{generates cifti specific manhattan plot}

\item{man.thresh}{visual horizontal bar for manhattan plot}

\item{niter}{number of iterations}

\item{orth}{determines whether orthogonal vectors are required}

\item{trace}{this allows debugging}

\item{v}{focuses on the loadings vector}

\item{center}{this centers the predictor columns}

\item{test_metrics}{flag for calculating test metrics}

\item{cnames}{this gives column names}

\item{vpos}{this determines whether the matrix is to be positive only}

\item{vneg}{this determines whether the matrix is to be negative only}

\item{compute.pve}{computes cv using missing values}

\item{not}{currently supported}

\item{nresp}{number of response vectors present in datasets}
}
\description{
This function provides all options for analysis using SSPCA
}
\examples{
OmniSPCA <- function(xtrain, ytrain, xtest, ytest, resp.names = NULL, 
  ycont=T, ybinary=F, sepAnalysis = F, 
  K=1, sumabsv=0, strictEV=T, Sens.analysis=F,
  sepclass=T, binClassCombine=F, NoExtraEV = F, ydensity=F,
  dpyhat=F, dsq=F, qqplot=F, opplot=F, 
  Evec.plot = F, manhattan.plot=F, cifti.man.plot=F, 
  man.thresh = 0, niter=20, orth=TRUE, 
  trace=TRUE, v=NULL, center=TRUE, test_metrics=T,
  cnames=NULL, vpos=FALSE, vneg=FALSE, compute.pve=TRUE)
}
