% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPplot.R
\name{OPplot}
\alias{OPplot}
\title{This function provides a comparison for the training and testing data
with the fitted values}
\usage{
OPplot(ytrain, ytest, yhat, nresp, resp.names)
}
\arguments{
\item{ytrain}{training response dataset}

\item{ytest}{testing response dataset}

\item{yhat}{this is the fitted data to be compared to the training and testing}

\item{nresp}{number of response vectors present in datasets}

\item{resp.names}{gives list of names for each response}
}
\description{
This function provides a comparison for the training and testing data
with the fitted values
}
\examples{
QQplot(ytrain, ytest, nresp, resp.names)
}
\keyword{observed}
\keyword{plot}
\keyword{predicted}
