% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermBinary.R
\name{PermBinary}
\alias{PermBinary}
\title{This function caclulates all possible response permutations for a series
of binary response variables}
\usage{
PermBinary(n)
}
\arguments{
\item{n}{the number of binary responses}
}
\description{
This function caclulates all possible response permutations for a series
of binary response variables
}
\examples{
PermBinary(n)
}
\keyword{binary}
\keyword{permutation}
