% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMDL1.R
\name{PMDL1}
\alias{PMDL1}
\title{This function supervises the decomposition of the matrix X into vectors
corresponding to the observations and sparse vectors corresponding to
the loadings}
\usage{
PMDL1(
  x,
  sumabs = 0.4,
  sumabsu = NULL,
  sumabsv = NULL,
  niter = 20,
  K = 1,
  v = NULL,
  trace = TRUE,
  orth = TRUE,
  center = TRUE,
  rnames = NULL,
  cnames = NULL,
  upos = upos,
  uneg = uneg,
  vpos = vpos,
  vneg = vneg
)
}
\arguments{
\item{x}{this is the matrix which is decomposed and the decomposition vectors
corresponding to the loadings are shrunk}

\item{sumabsu}{this parameter determines by how much the decomposition vectors
corresponding to the observations}

\item{sumabsv}{this parameter determines by how much the decomposition vectors
corresponding to the loadings}

\item{niter}{number of iterations to allow for convergence}

\item{K}{determines how many vectors should be calculated}

\item{v}{allows a particular v matrix to be passed to the MultiSMD and
MultiSMDOrth functions.  Its default is set to null because
CheckPMDV is called in this function}

\item{trace}{option to deisplay potential warnings}

\item{orth}{specifies that MultiSMDOrth is to be called for any number of
vectors that is greater than 1}

\item{center}{this tells the algorithm whether centering is required
this has been handled in other programs so has been disabled}

\item{rnames}{gives potential row names}

\item{cnames}{gives potential column names}

\item{upos}{allows the option to force the vector corresponding to
observations to correspond to the parallel maxima using pmax}

\item{uneg}{allows the option to force the vector corresponding to
observations to correspond to the parallel minima using pmin}

\item{vpos}{allows the option to force the vector corresponding to
loadings to correspond to the parallel maxima using pmax}

\item{vneg}{allows the option to force the vector corresponding to
loadings to correspond to the parallel minima using pmin}
}
\description{
This function supervises the decomposition of the matrix X into vectors
corresponding to the observations and sparse vectors corresponding to
the loadings
}
\examples{
PMDL1(x,sumabs=.4,sumabsu=NULL,sumabsv=NULL,niter=20,
  K=1,v=NULL, trace=TRUE, orth=TRUE, center=TRUE, rnames=NULL, cnames=NULL, 
  upos=upos, uneg=uneg, vpos=vpos, vneg=vneg)
}
\keyword{decomposition}
\keyword{matrix}
\keyword{partial}
