% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recPack.R
\name{recPack}
\alias{recPack}
\title{This function assigns a category based on binary response permutation}
\usage{
recPack(ytrain)
}
\arguments{
\item{ytrain}{accepts matrix of response variables}
}
\description{
This function assigns a category based on binary response permutation
}
\examples{
recPack(ytrain)
}
\keyword{reorder}
