% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructuring.func.R
\name{restructuring.func}
\alias{restructuring.func}
\title{This function converts the eigenecvectors into a new form.}
\usage{
restructuring.func(Z, resp.names, nresp)
}
\arguments{
\item{Z}{contains eigenvectors}

\item{resp.names}{gives list of names for each response}

\item{nresp}{number of response vectors present in datasets}
}
\description{
This function converts the eigenecvectors into a new form.
}
\examples{
restructuring.func(Z, resp.names, nresp)
}
\keyword{restructuring}
