% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safesvd.R
\name{safesvd}
\alias{safesvd}
\title{A function that performs SVD
This function iterates 10 times to attempt to decompose a matrix
through the singular value decomposition.  If the function is unable
to decompose the matrix after 10 tries, it will construct a random matrix
using rnorm with the same dimensions as the original matrix.  This random
matrix will be decomposed using the SVD and returned to the user.}
\usage{
safesvd(x)
}
\arguments{
\item{x}{This parameter accepts the matrix to undergo SVD}
}
\description{
A function that performs SVD
This function iterates 10 times to attempt to decompose a matrix
through the singular value decomposition.  If the function is unable
to decompose the matrix after 10 tries, it will construct a random matrix
using rnorm with the same dimensions as the original matrix.  This random
matrix will be decomposed using the SVD and returned to the user.
}
\examples{
safesvd(x)
}
\keyword{decomposition}
\keyword{singular}
\keyword{value}
