% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.R
\name{SMD}
\alias{SMD}
\title{This function decomposes a matrix using singular value decomposition.
The resulting eigenvector corresponding to the loadings in then shrunk
using an L1 penalty.}
\usage{
SMD(x, sumabsu, sumabsv, niter = 20, trace = TRUE, v, upos, uneg, vpos, vneg)
}
\arguments{
\item{x}{this is the matrix which is decomposed and the decomposition vectors
corresponding to the loadings are shrunk}

\item{sumabsu}{this parameter determines by how much the decomposition vectors
corresponding to the observations}

\item{sumabsv}{this parameter determines by how much the decomposition vectors
corresponding to the loadings}

\item{niter}{number of iterations to allow for convergence}

\item{trace}{option to deisplay potential warnings}

\item{v}{allows a particular loadings vector to be used}

\item{upos}{allows the option to force the vector corresponding to
observations to correspond to the parallel maxima using pmax}

\item{uneg}{allows the option to force the vector corresponding to
observations to correspond to the parallel minima using pmin}

\item{vpos}{allows the option to force the vector corresponding to
loadings to correspond to the parallel maxima using pmax}

\item{vneg}{allows the option to force the vector corresponding to
loadings to correspond to the parallel minima using pmin}
}
\description{
This function decomposes a matrix using singular value decomposition.
The resulting eigenvector corresponding to the loadings in then shrunk
using an L1 penalty.
}
\examples{
SMD(x, sumabsu, sumabsv, niter=20,trace=TRUE, v, upos, uneg, vpos, vneg)
}
\keyword{matrix}
\keyword{singular}
