% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMDOrth.R
\name{SMDOrth}
\alias{SMDOrth}
\title{This function decomposes a matrix using singular value decomposition but
constructs the new u vectors such that they are orthogonal to the previous
vectors
The resulting eigenvector corresponding to the loadings in then shrunk
using an L1 penalty where the u vectors are orthogonal to previous u
vectors.}
\usage{
SMDOrth(x, us, sumabsv = NULL, niter = 20, trace = TRUE, v, vpos, vneg)
}
\arguments{
\item{x}{this is the matrix which is decomposed and the decomposition vectors
corresponding to the loadings are shrunk}

\item{us}{this tells the us the previous u vectors for which we wish
to construct new u vectors that are orthogonal}

\item{sumabsv}{this parameter determines by how much the decomposition vectors
corresponding to the loadings}

\item{niter}{number of iterations to allow for convergence}

\item{trace}{option to deisplay potential warnings}

\item{v}{allows a particular loadings vector to be used}

\item{vpos}{allows the option to force the vector corresponding to
loadings to correspond to the parallel maxima using pmax}

\item{vneg}{allows the option to force the vector corresponding to
loadings to correspond to the parallel minima using pmin}
}
\description{
This function decomposes a matrix using singular value decomposition but
constructs the new u vectors such that they are orthogonal to the previous
vectors
The resulting eigenvector corresponding to the loadings in then shrunk
using an L1 penalty where the u vectors are orthogonal to previous u
vectors.
}
\examples{
SMDOrth(x, us, sumabsv=NULL, niter=20, trace=TRUE,v, vpos, vneg)
}
\keyword{matrix}
\keyword{orthogonal}
\keyword{singular}
