% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{A soft threshold function
This function shrinks a vector by a given amount}
\usage{
soft(x, d)
}
\arguments{
\item{x}{This parameter accepts a vector to be shrunk}

\item{d}{This parameter represents the magnitude fo the shrinkage on vector x}
}
\description{
A soft threshold function
This function shrinks a vector by a given amount
}
\examples{
soft(x,d)
}
\keyword{soft}
\keyword{threshold}
