% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPC.R
\name{SPC}
\alias{SPC}
\title{This function computes supervised principal component analysis.  It obtains
the eigenevectors from the HSIC criterion and encodes the training and
testing data}
\usage{
SPC(ytrain, xtrain, ytest, xtest, ycont, ybinary, nresp, strictEV, nTopEvecs)
}
\arguments{
\item{ytrain}{training response dataset}

\item{xtrain}{training predictor dataset}

\item{ytest}{testing response dataset}

\item{xtest}{testing predictor dataset}

\item{ycont}{flag indicates response data is continuous}

\item{ybinary}{flag indicates response data is binary}

\item{nresp}{number of response vectors present in datasets}

\item{strictEV}{flag whether to impose strict requiresments to accept number
of eigenevectors or to select a more appropriate number}

\item{nTopEvecs}{number of desired eigenevectors}
}
\description{
This function computes supervised principal component analysis.  It obtains
the eigenevectors from the HSIC criterion and encodes the training and
testing data
}
\examples{
SPC(ytrain, xtrain, ytest, xtest, ycont, ybinary, nresp, strictEV, nTopEvecs)
}
\keyword{analysis}
\keyword{component}
\keyword{principal}
\keyword{supervised}
